some file restructuring
