+++ data-science/cnn/model_av/card_classifier_a.py
