+++ data-science/testin-er-out.py
@@ -20,7 +20,7 @@ from_PIL = transform.Compose([transform.PILToTensor()])
 # img = np.array(Image.open(BytesIO(resp.content)))
 
 # get and load up card from file
-img = np.array(Image.open('data-science/data/card.jpg'))
+img = np.array(Image.open('data-science/data/collected/card.jpg'))
 
 # rearrange color channels >>> RGBA to BGR, RGB to BGR, etc...
 if img.shape[2] == 4:
